Hatching creates correct rim, no splitting of the inner rings.
However, the inner rings are not yet processed, nor is there a
variable structure defined to track what was processed yet,
nor is there a way to find out what are "front" and "rear" segments.
It must be noted that the n-object has far too much size, as each one brings
actual multilistreference's. This multilistreference stuff must change
somehow, in a sense that allows for leaner "empty" objects.
Depth is not defined for the hatches, nor are they actual objects.
Speaking of, I intend to have all not contained objects be referenced by id.
AND: I added a rim over the spheres, to prevent a cut-off of points lying directly on the rim.

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>