SDL shows picture
Split of SVG and SDL output occurs by macro. If both should be used,
I must implement the corresponding part of createsvg twice, and differently.
I have to draw out the svg writes from the >>controlprocedure<< and use
macro-defined different names, and the switch should be done by self-
modifying code. Or I will have to use "graphics objects" or define the corresponding
procedures dynamically. Better: make the controlprocedure a part of the
graphics object's main definition,
but implemented in a different unit that is used twice.

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>