Bondorder-depending hydrogen count, editing bondorder, further atomsymbol display
So that an atom does not need to be changed when a bond is added,
the hydrogen count is stored as how many would be there if no bond was
attached.

Fullscreen works now. One's gotta be careful with the released key interpretation.
Most key interpretation in the main loop forks

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>