Make multilists rely solely
TODO: restore multilist pointer after undo (only in restored multilist)
Shift pointers in undobuffer on memory motions#
or is it better to have the undo-buffer just a handle to the memory,
an index into the memory list?
There will be no need to transfer the memory adressadress to the undo buffer,
because each undo buffer is a new copy, and also gets restored from by copy.

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>