Memory management, sort improved, faster
using a memory management routine capable of scaling

Warning! I did not apply the malloc for all purposes, just for undo buffers!
and there not for the main undo buffer.
Warning! no minimum size checking employed for the half-memory tool!

janitor's sort stuff has to be improved further.
Especially, since I want a Sort-into, using the spaces in the sort array,
which should allow for merger of the old and a Z-numbered set of new elements.

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>